refactor: eventHandler디렉토리 내 파일 분할
